+++ ntex-io/CHANGES.md
@@ -1,5 +1,9 @@
 # Changes
 
+## [0.3.1] - 2023-06-23
+
+* `PipelineCall` is static
+
 ## [0.3.0] - 2023-06-22
 
 * Release v0.3.0